backend/pyproject.toml
[project]
name = "cnc-iot-backend"
version = "0.1.0"
description = "CNC IoT Backend API"
requires-python = "==3.12.*"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.29.0",
    "sqlalchemy==2.0.30",
    "psycopg2-binary==2.9.9",
    "pydantic==2.7.1",
    "pydantic-settings==2.2.1",
    "python-dotenv==1.0.1",
    "httpx",
    "pytest",
    "pytest-asyncio",
    "ruff",
    "requests>=2.33.1",
    "customtkinter>=5.2.2",
    "paho-mqtt>=2.1.0",
    "influxdb-client>=1.43.0",
]

[project.optional-dependencies]
dev = [
    "httpx",
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
